+++ cookbooks/base/recipes/main-debian.rb
@@ -1,7 +1,7 @@
 #
 # Author:: Carles Muiños (<carles.ml.dev@gmail.com>)
 # Cookbook Name:: base
-# Recipe:: main-linuxmint
+# Recipe:: main-debian
 #
 # Copyright 2013,2014 Carles Muiños
 #
@@ -20,56 +20,21 @@
 
 
 # Language support
-package "language-pack-ca"
-package "language-pack-gnome-ca"
 package "aspell-ca"
 package "myspell-ca"
 
-package "language-pack-es"
-package "language-pack-gnome-es"
 package "aspell-es"
 package "myspell-es"
 
 
-os_release = platform_version
-
-# Support for ia32-libs dependency
-if platform_arch == "x86_64" and os_release == 13
-  %w[
-    ia32-libs-multiarch
-    i386
-    lib32gcc1
-    libc6-i386
-  ].each do |pkg|
-    package pkg
-  end
-end
-
-
 # Virtualization support
-unless virtual_box?
+if virtual_box?
   %w[
     virtualbox-guest-dkms
     virtualbox-guest-utils
     virtualbox-guest-x11
   ].each do |pkg|
-    package pkg do
-      action :purge
-    end
-  end
-end
-
-
-# Commonly used restricted packages for Linux Mint
-if os_release < 15
-  repo_name = "videolan-#{platform_codename}"
-
-  apt_repository repo_name do
-    uri "http://download.videolan.org/pub/debian/stable/ /"
-    distribution ""
-    key "http://download.videolan.org/pub/debian/videolan-apt.asc"
-    action :add
-    not_if { ::File.exists? "#{node[:apt][:sources_path]}/#{repo_name}.list" }
+    package pkg
   end
 end